fixed narrowing conersion bug

+++ Library/Sources/Stroika/Foundation/Execution/Platform/Windows/Users.cpp
@@ -31,9 +31,8 @@ String  Platform::Windows::SID22UserName (PSID sid)
     SID_NAME_USE    iUse {};
     SDKChar         name[1024];
     SDKChar         domain[1024];
-    DWORD           nameLen     { NEltsOf (name) };
-    DWORD           domainLen   { NEltsOf (domain) };
-
+    DWORD           nameLen     { static_cast<DWORD> (NEltsOf (name)) };
+    DWORD           domainLen   { static_cast<DWORD> (NEltsOf (domain)) };
     ThrowIfFalseGetLastError (::LookupAccountSid (nullptr, sid, name, &nameLen, domain, &domainLen, &iUse));
     if (domainLen == 0) {
         return String::FromSDKString (name);
